\name{SpikeIn}
\alias{SpikeIn133}
\alias{SpikeIn95}
\title{Affymetrix SpikeIn Experiment Data}
\description{
  Probe-level data for data from the HGU95 and HGU133 SpikeIn experiments. 
}
\usage{
data(SpikeIn95)
data(SpikeIn133)
}
\format{\code{SpikeIn} is \code{\link[affy:ProbeSet-class]{ProbeSet}}
  containing the 
  $PM$ and $MM$ intensities for a genes spiked in at different
  concentrations. To see the spike-in concentrations use \code{pData}.}
\source{
For more information see Irizarry, R.A., et al. NAR (2003)
\url{http://www.biostat.jhsph.edu/~ririzarr/papers/index.html}}

\keyword{datasets}
